requirements.txt
uvicorn
psycopg2-binary
fastapi
starlette
